rename graph_based_model_construction to model_construction
